nextjs file copy
