+++ src-tauri/Cargo.toml
@@ -17,7 +17,7 @@ tauri-build = { version = "1.2.1", features = [] }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.2.4", features = ["fs-all", "path-all", "shell-execute", "shell-open", "shell-sidecar"] }
+tauri = { version = "1.2.4", features = ["fs-all", "path-all", "shell-all"] }
 tokio = { version = "1.15.0", features = ["full"] }
 serialport = "4.2.0"
 hex = "0.4.3"
