+++ src-tauri/tauri.conf.json
@@ -1,10 +1,11 @@
 {
   "$schema": "../node_modules/@tauri-apps/cli/schema.json",
   "build": {
-    "beforeBuildCommand": "npm run build",
+    "beforeBuildCommand": "npm run build && npm run export",
     "beforeDevCommand": "npm run dev",
-    "devPath": "http://localhost:5173",
-    "distDir": "../dist"
+    "devPath": "http://localhost:3000",
+    "distDir": "../out",
+    "withGlobalTauri": true
   },
   "package": {
     "productName": "key.board",
@@ -19,6 +20,7 @@
         "all": true
       },
       "shell": {
+        "all": true,
         "execute": true,
         "sidecar": true,
         "open": true,
